@@ -11,7 +11,7 @@
 	"issues": "https://omnia.nexx.cloud/support"
   },
   "require": {
-	"php": "^7.4 || ^8.0 || ^8.1  || ^8.2",
+	"php": "^8.0 || ^8.1  || ^8.2",
 	"ext-json": "*",
 	"guzzlehttp/guzzle": "6.*|7.*",
 	"microsoft/azure-storage-blob": "1.*",
